[GSCOLLECT-1445] Fix invalid Java that javac accepts but is invalid according to the spec.
git-svn-id: svn+ssh://gscollections.svn.services.gs.com/svnroot/gscollections-svn/trunk@680 d5c9223b-1aff-41ac-aadd-f810b4a99ac4

+++ collections/src/main/java/com/gs/collections/impl/bag/sorted/mutable/TreeBag.java
@@ -1239,7 +1239,7 @@ public class TreeBag<T>
             public void value(T each, int occurrences)
             {
                 long longValue = function.longValueOf(each);
-                sum[0] += longValue * (long) occurrences;
+                sum[0] += longValue * occurrences;
             }
         });
         return sum[0];
@@ -1254,7 +1254,7 @@ public class TreeBag<T>
             public void value(T each, int occurrences)
             {
                 double doubleValue = function.doubleValueOf(each);
-                sum[0] += doubleValue * (double) occurrences;
+                sum[0] += doubleValue * occurrences;
             }
         });
         return sum[0];
